requirements.txt
nltk
lru-dict